Changed PID, uploaded SLAM logs
Changed PID values within cf2_defaults, and uploaded a log of the SLAM data as a result. Worked a ton better, lots of progress. System still oscillates a ton, and I think the position gains are probably what needs the most changes. Look into increasing the damping and/or integrating term. We might be okay to remove the damping term altogether.